+++ examples/factory_async.rs
@@ -109,7 +109,7 @@ impl AsyncFactoryComponent for Counter {
         Some(LoadingWidgets::new(root, spinner))
     }
 
-    fn output_to_parent_input(output: Self::Output) -> Option<AppMsg> {
+    fn forward_to_parent(output: Self::Output) -> Option<AppMsg> {
         Some(match output {
             CounterOutput::SendFront(index) => AppMsg::SendFront(index),
             CounterOutput::MoveUp(index) => AppMsg::MoveUp(index),