+++ relm4-macros/src/visitors.rs
@@ -4,9 +4,11 @@ use proc_macro2::TokenStream;
 use quote::quote;
 use syn::visit::{self, Visit};
 use syn::visit_mut::{self, VisitMut};
+use syn::LocalInit;
 
 use crate::additional_fields::AdditionalFields;
 use crate::menu::Menus;
+use crate::util;
 use crate::widgets::ViewWidgets;
 
 #[derive(Debug)]
@@ -91,10 +93,10 @@ impl VisitMut for ComponentVisitor<'_> {
                     _ => (),
                 }
             }
-            syn::ImplItem::Method(func) => {
+            syn::ImplItem::Fn(func) => {
                 if &*func.sig.ident.to_string() == "init" {
                     let mut init_fn_visitor = InitFnVisitor::default();
-                    init_fn_visitor.visit_impl_item_method(func);
+                    init_fn_visitor.visit_impl_item_fn(func);
 
                     self.model_name = init_fn_visitor.model_name;
                     self.sender_name = init_fn_visitor.sender_name;
@@ -123,7 +125,8 @@ impl VisitMut for ComponentVisitor<'_> {
 pub(super) struct FactoryComponentVisitor<'errors> {
     pub(super) view_widgets: Option<syn::Result<ViewWidgets>>,
     pub(super) widgets_ty: Option<syn::Type>,
-    pub(super) init_widgets: Option<syn::ImplItemMethod>,
+    pub(super) index_ty: Option<syn::Type>,
+    pub(super) init_widgets: Option<syn::ImplItemFn>,
     pub(super) root_name: Option<syn::Ident>,
     pub(super) additional_fields: Option<AdditionalFields>,
     pub(super) menus: Option<Menus>,
@@ -135,6 +138,7 @@ impl<'errors> FactoryComponentVisitor<'errors> {
         FactoryComponentVisitor {
             view_widgets: None,
             widgets_ty: None,
+            index_ty: None,
             init_widgets: None,
             root_name: None,
             additional_fields: None,
@@ -199,10 +203,10 @@ impl VisitMut for FactoryComponentVisitor<'_> {
                     _ => (),
                 }
             }
-            syn::ImplItem::Method(func) => {
+            syn::ImplItem::Fn(func) => {
                 if &*func.sig.ident.to_string() == "init_widgets" {
                     let mut init_fn_visitor = InitWidgetsFnVisitor::default();
-                    init_fn_visitor.visit_impl_item_method(func);
+                    init_fn_visitor.visit_impl_item_fn(func);
 
                     self.root_name = init_fn_visitor.root_name;
                     self.errors.append(&mut init_fn_visitor.errors);
@@ -234,6 +238,8 @@ impl VisitMut for FactoryComponentVisitor<'_> {
                 ty,
                 "`Root` type is defined by `view!` macro",
             ));
+        } else if ty.ident == "Index" {
+            self.index_ty = Some(ty.ty.clone());
         }
     }
 }
@@ -247,54 +253,24 @@ struct InitFnVisitor {
 }
 
 impl<'ast> Visit<'ast> for InitFnVisitor {
-    fn visit_impl_item_method(&mut self, func: &'ast syn::ImplItemMethod) {
-        let root_name = match func.sig.inputs.iter().nth(1) {
-            Some(syn::FnArg::Typed(pat_type)) => match &*pat_type.pat {
-                syn::Pat::Ident(ident) => Ok(ident.ident.clone()),
-                _ => Err(syn::Error::new_spanned(
-                    pat_type,
-                    "unable to determine root name",
-                )),
-            },
-            Some(arg) => Err(syn::Error::new_spanned(
-                arg,
-                "unable to determine root name",
-            )),
-            None => Err(syn::Error::new_spanned(
-                &func.sig,
-                "unable to determine root name",
-            )),
-        };
+    fn visit_impl_item_fn(&mut self, func: &'ast syn::ImplItemFn) {
+        let Some(root_arg) = func.sig.inputs.iter().nth(1) else { return };
+        let root_name = util::extract_arg_ident(root_arg);
 
         match root_name {
-            Ok(root_name) => self.root_name = Some(root_name),
+            Ok(root_name) => self.root_name = Some(root_name.clone()),
             Err(e) => self.errors.push(e),
         }
 
-        let sender_name = match func.sig.inputs.iter().nth(2) {
-            Some(syn::FnArg::Typed(pat_type)) => match &*pat_type.pat {
-                syn::Pat::Ident(ident) => Ok(ident.ident.clone()),
-                _ => Err(syn::Error::new_spanned(
-                    pat_type,
-                    "unable to determine sender name",
-                )),
-            },
-            Some(arg) => Err(syn::Error::new_spanned(
-                arg,
-                "unable to determine sender name",
-            )),
-            None => Err(syn::Error::new_spanned(
-                &func.sig,
-                "unable to determine sender name",
-            )),
-        };
+        let Some(sender_arg) = func.sig.inputs.iter().nth(2) else { return };
+        let sender_name = util::extract_arg_ident(sender_arg);
 
         match sender_name {
-            Ok(sender_name) => self.sender_name = Some(sender_name),
+            Ok(sender_name) => self.sender_name = Some(sender_name.clone()),
             Err(e) => self.errors.push(e),
         }
 
-        visit::visit_impl_item_method(self, func);
+        visit::visit_impl_item_fn(self, func);
     }
 
     fn visit_expr_struct(&mut self, expr_struct: &'ast syn::ExprStruct) {
@@ -343,31 +319,16 @@ struct InitWidgetsFnVisitor {
 }
 
 impl<'ast> Visit<'ast> for InitWidgetsFnVisitor {
-    fn visit_impl_item_method(&mut self, func: &'ast syn::ImplItemMethod) {
-        let root_name = match func.sig.inputs.iter().nth(2) {
-            Some(syn::FnArg::Typed(pat_type)) => match &*pat_type.pat {
-                syn::Pat::Ident(ident) => Ok(ident.ident.clone()),
-                _ => Err(syn::Error::new_spanned(
-                    pat_type,
-                    "unable to determine root name",
-                )),
-            },
-            Some(arg) => Err(syn::Error::new_spanned(
-                arg,
-                "unable to determine root name",
-            )),
-            None => Err(syn::Error::new_spanned(
-                &func.sig,
-                "unable to determine root name",
-            )),
-        };
+    fn visit_impl_item_fn(&mut self, func: &'ast syn::ImplItemFn) {
+        let Some(root_arg) = func.sig.inputs.iter().nth(2) else { return };
+        let root_name = util::extract_arg_ident(root_arg);
 
         match root_name {
-            Ok(root_name) => self.root_name = Some(root_name),
+            Ok(root_name) => self.root_name = Some(root_name.clone()),
             Err(e) => self.errors.push(e),
         }
 
-        visit::visit_impl_item_method(self, func);
+        visit::visit_impl_item_fn(self, func);
     }
 }
 
@@ -394,7 +355,7 @@ impl<'errors> PreAndPostView<'errors> {
 
 impl VisitMut for PreAndPostView<'_> {
     fn visit_impl_item_mut(&mut self, item: &mut syn::ImplItem) {
-        if let syn::ImplItem::Method(func) = item {
+        if let syn::ImplItem::Fn(func) = item {
             match &*func.sig.ident.to_string() {
                 "pre_view" => {
                     if !self.pre_view.is_empty() {
@@ -459,9 +420,9 @@ impl ViewOutputExpander<'_> {
 }
 
 impl VisitMut for ViewOutputExpander<'_> {
-    fn visit_impl_item_method_mut(&mut self, method: &mut syn::ImplItemMethod) {
+    fn visit_impl_item_fn_mut(&mut self, method: &mut syn::ImplItemFn) {
         if method.sig.ident == "init" || method.sig.ident == "init_widgets" {
-            visit_mut::visit_impl_item_method_mut(self, method);
+            visit_mut::visit_impl_item_fn_mut(self, method);
 
             if !self.expanded {
                 self.errors.push(syn::Error::new_spanned(method, "expected an injection point for the view macro. Try using `let widgets = view_output!();`"));
@@ -473,7 +434,7 @@ impl VisitMut for ViewOutputExpander<'_> {
         let mut expand = false;
 
         if let syn::Stmt::Local(syn::Local {
-            init: Some((_, expr)),
+            init: Some(LocalInit { expr, .. }),
             ..
         }) = stmt
         {
@@ -493,10 +454,13 @@ impl VisitMut for ViewOutputExpander<'_> {
         if expand {
             let view_code = &self.view_code;
 
-            *stmt = syn::Stmt::Expr(syn::Expr::Verbatim(quote! {
-                #view_code
-                #stmt
-            }));
+            *stmt = syn::Stmt::Expr(
+                syn::Expr::Verbatim(quote! {
+                    #view_code
+                    #stmt
+                }),
+                None,
+            );
 
             self.expanded = true;
         }