+++ relm4/src/component/async/traits.rs
@@ -27,7 +27,7 @@ pub trait AsyncComponent: Sized + 'static {
     /// The parameter used to initialize the component.
     type Init;
 
-    /// The widget that was constructed by the component.
+    /// The top-level widget of the component.
     type Root: Debug + Clone;
 
     /// The type that's used for storing widgets created for this component.
@@ -160,7 +160,7 @@ pub trait SimpleAsyncComponent: Sized + 'static {
     /// The parameter used to initialize the component.
     type Init;
 
-    /// The widget that was constructed by the component.
+    /// The top-level widget of the component.
     type Root: Debug + Clone;
 
     /// The type that's used for storing widgets created for this component.