+++ relm4/src/extensions/set_child.rs
@@ -41,7 +41,8 @@ set_child_impl!(
     gtk::Overlay,
     gtk::Revealer,
     gtk::WindowHandle,
-    gtk::Expander
+    gtk::Expander,
+    gtk::AspectFrame
 );
 
 #[cfg(feature = "libadwaita")]