+++ relm4/src/factory/sync/mod.rs
@@ -7,5 +7,5 @@ mod traits;
 use builder::FactoryBuilder;
 use handle::FactoryHandle;
 
-pub use collections::{FactoryVecDeque, FactoryVecDequeGuard};
+pub use collections::{FactoryHashMap, FactoryVecDeque, FactoryVecDequeGuard};
 pub use traits::{CloneableFactoryComponent, FactoryComponent};