+++ CHANGES.md
@@ -5,15 +5,29 @@
 ### Added
 
 + core: Add `TypedListView` as idiomatic wrapper over `gtk::ListView`
+
+### Fixed
+
++ macros: Improve error messages for non-identifier parameter patterns
+
+## 0.6.0-beta.1 - 2023-4-19
+
++ core: Introduce setting and action safeties
++ core: Implement `RelmSetChildExt` for `gtk::AspectFrame`
++ core: Add `FactoryHashMap` as alternative to `FactoryVecDeque`
 + core: Add gnome_44 feature flag for GNOME 44
 + core: Documentation and better support for data bindings
 + core: Add `set_tooltip` method to `RelmWidgetExt`
 + core: Add `main_adw_application` method to retrieve the `adw::Application` when the libadwaita feature is enabled
 + macros: Add `skip_macro` option for watch and track attributes to skip their initialization 
++ examples: Introduce setting and action safeties
 + examples: Example for using relm4-icons 
 
 ### Changed
 
++ core: Replace `FactoryVecDeque`'s associated function `from_vec` with `from_iter`
++ core: Added `Index` type to the `FactoryComponent` trait
++ core: Rename factory component traits `output_to_parent_input` method to `forward_to_parent`
 + core: Improved `RelmActionGroup` API
 + all: Increase MSRV to 1.65 to match the dependencies